requirements.txt
Flask
mysql.connector
server.py

